pyproject.toml
[project]
name = "mlops-xoxo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.63.0",
    "mlflow<3.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "torch",
    "torchvision",
    "facenet-pytorch",
    "mlflow",
    "dvc",
    "opencv-python",
    "scikit-learn",
    "matplotlib",
    "pandas",
    "tqdm",
    "pillow",
    "seaborn",
    "pyarrow",
    "pyyaml",
    "dotenv",
    "codecarbon>=3.0.5",
    "pylint>=3.3.8",
    "flake8>=7.3.0",
    "pytest-mock>=3.15.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pylint>=3.3.8",
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
